Fix grit unittests.
The two rc unittests are failing because it appears the data
files checked into svn don't match.  It's easier to change the
tests because the data is in a different (internal only) repo.
AFAICT, this has been broken forever.

The other fix is because a test became more correct after
r43930, but the test was passing anyway.

Review URL: http://codereview.chromium.org/2131013