Fix a ChromeFrame crash caused when the cookie policy is set to prompt and the active document is released before the prompt is clicked on. This only happens on IE6 where the prompt dialog is modeless in the context of the InternetSetCookieEx call.
To ensure that the active document remains valid for the duration of the call we maintain a reference on the container for the duration of
the InternetSetCookie call.

The cookie calls in IE are now handled on the UI thread as we cannot block the IPC thread in case the set cookie calls result in a prompt.

This fixes bug http://code.google.com/p/chromium/issues/detail?id=44465

The FullTabModeIE_UnloadEventTest has been disabled as the change to execute the cookie calls on the UI thread would break it. 

Bug=44465, 40814

Review URL: http://codereview.chromium.org/2091015