Buildbot and BVT integration changes.
I've purposely left the 'bvt' step out of the x86 generic full builder for right now. I.e. none of this code will be executed. 'bvt' is also listed as a forgiving step. Once all the pieces are in place I'll add 'bvt' step and we'll be ready to roll. The forgiving step status will be evaluated after a burn in period.

Per Marc-Antoine, the wait_for_bvt.sh script will go in with a separate change set on the internal SVN.

Review URL: http://codereview.chromium.org/2124014