Fix a regression on Windows introduced by r48650.
For some reason, passing NULL as the hCertStore argument
to CertAddEncodedCertificateToStore causes CertGetCertificateChain
to fail to build the certificate chain for some servers,
even though the hCertStore member of the server certificate
context should contain its CA certificates.

So I reverted to passing our in-memory certificate store
as the hCertStore argument to CertAddEncodedCertificateToStore.

R=agl,mark
BUG=45706
TEST=Visit https://moversguide.usps.com/ on Windows.  Should
not get the "certificate is not trusted" SSL error page.
Review URL: http://codereview.chromium.org/2605007