On the 375 branch, change mini_installer to handle BLOB_ files.
The padded installers we are using for a download size experiment include
BLOB_6MB.TXT, BLOB_9MB.TXT, etc. payloads. Currently, the installer fails
after unpacking this part of the payload.

Note that this is the same as http://codereview.chromium.org/2086009, which 
worked on the 249 branch. I don't want to pollute trunk with this code that
will be obsolete at the end of the experiment.

BUG= None
TEST= mini_installer_9mb.exe works
R= kuchhal, cpu

Review URL: http://codereview.chromium.org/2776001