Fix and re-enable RedirectTest.ClientCancelled.
The hack to create and dispatch an event after a timeout to fool the browser
into thinking it is user initiated stopped working after WebKit r49827.

Changing the test to navigate to a javascript: url instead, which qualifies
as a user initiated location change.

BUG=45053
TEST=ui_tests --gtest_filter=RedirectTest.ClientCancelled

Review URL: http://codereview.chromium.org/2225007