linux: work around gnome keyring memory corruption
Older versions of gnome keyring corrupt memory if you fetch too much
data from it.  Despite our best efforts, we couldn't come up with a way
to tell if you're using the broken old version so that we could blacklist
it.  Instead, we fetch passwords one-by-one, which is slow but tolerable
since we do it on a background thread anyway.

While I'm here, refactor the code a bit to simplify it.

TEST=password manager doesn't crash when using mdm's sample 100 password db
with --password-store=gnome on hardy

Review URL: http://codereview.chromium.org/2838023