Add explicit file rename/cp information to svn rietveld uploads
Right now if you svn cp or svn mv before uploading to rietveld,
there's no description in the patch at all of what was done; so you
get a diff, but it's for who knows what revision of what prior file
to some current version.

For code reviews this kinda works (you ask the guy what it was), but for
trying to apply patches (ala git patch) this fails badly.  This patch
tries to add some metadata to the start of a rietveld patch that
describes these changes: both to a human (who can read them)
and to a potential clever future git-cl (which I'll do next).

The metadata looks like this after checking out a test repo, and
svn cp -r 1 foo nitz; svn cp bar quux; echo be good>>quux

  ### BEGIN SVN COPY METADATA
  #$ svn cp -r 1 foo nitz ### WARNING: note non-trunk copy
  #$ cp bar quux
  ### END SVN COPY METADATA
  Index: nitz
  Index: quux
  ===================================================================
  --- quux	(revision 0)
  +++ quux	(working copy)
  @@ -1,2 +1,3 @@
   hi mom
   hi sister
  +be good

I did a test, and this looks like it works in svn 1.4.4 as well.

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/2824035