Split Gaia auth out into common/net. This is in preparation for creating a new Auth service which will issue auth tokens.
I'd like to get it in and continue hacking on it so I can stop colliding so badly with merges.

I've ported over the existing unit tests. Eventually gaia_authenticator.cc is going away. For the time being, i'll be number two.

This is now done, with a pile of unit tests. I am trying to run the interactive tests but I was getting numerous unrelated failures and need to resync ChromiumOS to a working build.

TEST=unit tests, manually log in, out, enter wrong password.
BUG=47091

Review URL: http://codereview.chromium.org/2801015