Allow data_packages outputs to include flattened items without rc_all outputs.
This fixes a bug where the code to flatten an html file was in
the rc_all output step.  So if a data_package output came before
the rc_all output in the grd file, we would get a build error.

Just refactor the code to flatten for both output types (but only
flatten once).

BUG=48253
TEST=build completes succesfully

Review URL: http://codereview.chromium.org/2889001