Mac: Fail early on Sandbox errors
We know that on certain machines the Seatbelt library is borked and even a simple test program that initializes the sandbox fails.

There is already a CHECK() in chrome_dll_main() that fires if the Sandbox didn't initialize correctly.  This CL changes all sandbox failures to LOG(FATAL)s so we get different stack signatures when we crash.

BUG=44761
TEST=Browser & Unit tests should continue to function correctly.

Review URL: http://codereview.chromium.org/2861046