gclient: update git-cl whenever git-cl is run and not when gclient is run
Rather than updating git-cl whenever gclient is invoked, update it
whenever git-cl is invoked.

In the chromium-os workflow, folks often run git-cl but rarely run gclient.
This results in folks running an older git-cl which may not
have the latest features/fixes.

TEST=

Tested that git-cl is updated when it is stale, git-cl is cloned when it doesn't exist and is not updated when DEPOT_TOOLS_UPDATE=0

Review URL: http://codereview.chromium.org/2962013