Clean up FTP code a bit:
- remove unneeded states
- minor style and error handling improvements

This change removes ACCT command support. It was really untested, and I think it wouldn't work in case it was needed. We may add it back when we get a good test case. However, people are unlikely to use a web browser to access servers that require usage of ACCT command.

TEST=net_unittests
BUG=none

Review URL: http://codereview.chromium.org/2985004