Mac: Context Menu broken when tab dragged out to new window.
The cause of this bug was that in new windows created by dragging a tab out, the content view was ending up in front of the TabStripView and eating it's right-clicks.

The reason this was happening has to do with the order that the views were reconstitued in the new window in [TabWindowController moveViewsBetweenWindowAndOverlay].

This CL also fixes a few minor spelling issues seen in passing.

BUG=44023
TEST=Drag a tab out into a new Window, right click on the bottom half of the newly dragged out tab.  Context menu should open.  Previously such a click would do nothing.  See comment #10 on the bug for a screenshot of the area that used to be problematic.

Review URL: http://codereview.chromium.org/2918009