In ChromeFrame if window.open is received for a pending external tab, i.e. which has not received a connect from the host browser, currently we end up crashing chrome due to dereferencing a NULL automation channel.
Ideally we should queue all IPC messages destined for a pending tab and send them out once we receive a connect.
However that solution is tricky as the messages carry the tab handle as one of the arguments which is only available
when we receive a connnect for the tab. The AutomationHandleTracker gives out these handles and is per automation provider.

We may have different automation provider instances in the case of IE8 and hence we cannot pre compute the tab handle.

Current solution is to add a NULL check around the offending code path which would basically drop the new contents
notification.

Fixes bug http://code.google.com/p/chromium/issues/detail?id=49188

Bug=49188

Review URL: http://codereview.chromium.org/3021002