Revert 53603 - Make the OmniboxSearchHint infobar use PAGE_ACTION_TYPE.
With this, this infobar gets the blue style to match with the mocks.
Also did a cleanup to get rid of INFO_TYPE that is not more needed.

[Rollback: Following this patch, the Chromium XP debug builder started
having persistent link failures (LINK1318 error). This is a test to see
if it fixes the problems.]

BUG=39102
TEST=trybots

Signed-off-by: Thiago Farina <tfarina@chromium.org>

Review URL: http://codereview.chromium.org/2805096

TBR=tfarina@chromium.org
Review URL: http://codereview.chromium.org/2884038