Revert 53849 - Implement most of the startup page controls in DOMUI prefs
Hooks up display of startup pages, removal of pages, and setting the startup pages to the currently open tabs.
Uses a select for now, but will eventually be replaced with a List so the favicons can be displayed.

BUG=48713
TEST=Startup pages can be viewed, removed, and set to current in DOMUI prefs.

Review URL: http://codereview.chromium.org/3046025

TBR=stuartmorgan@chromium.org
Review URL: http://codereview.chromium.org/3058018