Fix glitch in rendering of the stroke at the bottom of the browser toolbar.
The code that rendered this line was doing weird things (adding/removing kClientEdgeThickness from the edges?!) so I removed this and used toolbar_bounds.right() instead of toolbar_bounds.width() which is correct given that toolbar_bounds.x() > 0.

I also got rid of the use of MirroredLeftPointForRect since the flipped version of a horizontal line is... the same horizontal line.

I have also updated the glass frame not to subtract a pixel from the right edge which looks glitchy as well.

http://crbug.com/48134
TEST=see bug

Review URL: http://codereview.chromium.org/3056022