Fix a bug where it's not possible to get out of fullscreen mode.
The bug happened when the previous state was maximized and
the monitor has no gnome-panels.  When unfullscreen'ing, the
maximized sized == the monitor size so metacity re-fullscreened.

I also fixed a bug where we assumed that screen size == monitor
size.  There's still a bug in SetWindowSize-- I'll fix that in
a follow up patch.

BUG=32347

Review URL: http://codereview.chromium.org/3077012