Merge 54331 - When running as a remote, call gdk_notify_startup_complete
Linux programs are expected to notify their launcher when they finish loading.
GTK+ does this automatically when the first GtkWindow is opened. When acting as
a remote, no window is opened, so we should manually send the notification.

This could perhaps be more thorough and pass DESKTOP_LAUNCH_ID to the remote to
call with gdk_notify_startup_complete_with_id, but a quick survey on codesearch
suggests that no one bothers anyway.

(This is reproducible in KDE. The bug report suggests some configurations of
GNOME need this as well, but not yet reproducible.)

R=estade
BUG=45392
TEST=throbber disappears on second Chrome launch in KDE

Review URL: http://codereview.chromium.org/3051023

TBR=davidben@chromium.org
Review URL: http://codereview.chromium.org/3077018