GTK: minor MenuBarHelper fix.
If one of the buttons is not realized, gtk_widget_translate_coordinates() will fail on it. Instead of aborting, just skip to the next button. This can happen when opening a new window with a bookmark bar that doesn't have an overflow chevron (when the chevron has not been shown, it will be unrealized).

BUG=51810
TESt=see bug + above description

Review URL: http://codereview.chromium.org/3175006