Run tests faster.
I noticed that net_unittests was spending a lot of time doing nothing.
This CL changes the timeouts so that we can startup and shutdown test
servers faster.

(This is a reland. Originally r54579, reverted in r54592. This is
relanding unmodified as I don't believe that it was mistakenly
fingered for build issues at the time. The real problem was r54576,
reverted in r54595.)

BUG=none
TEST=net_unittests

http://codereview.chromium.org/3057021/show