Clean up PLT histograms for a move to navigator_state.cc
Clean up PLT gathering in preparation for moving it
to take place in navigation_state.cc, during the
instance destructor.  This CL should have almost
no semantic impact, and I'm submitting it separately
so that when I move large blocks, I can mostly move
them unchanged.

Add a bunch of DCHECKs to be sure that finish and finish_doc
times are only set once, and that they are set only in
the correct order.

In a future CL, I expect to push a lot of the time-gathering
code into navigation_state.cc, along with the generation
of the final histograms.  The challenge is to mark the
NavigationState instances that are worthy of recording
(which looks like it MAY be done by marking the state
with the boolean to indicate it is part of an HTTP or
HTTPS connection), and make sure we have good numbers 
throughout.

bug=48970
r=mbelshe
Review URL: http://codereview.chromium.org/2993004