Landing r56438 to the M6 branch. Original review: http://codereview.chromium.org/3158017
Every time a new autofill suggestion was received we would hide the autofill popup before showing a new one. 
This causes it to flash every time the user type a new key (as we get new suggestions for each key stroke). 
It is safe to hide the autofill only when there are no suggestions. 
If there are suggestions, the WebKit side takes care of resizing and potentially hiding the popup if necessary. 

BUG=49980
TEST=Trigger the AutoFill/Autocomplete popup. As you type characters in the text input, the popup should not flash. Test that it does not flash either when you press backspace.
TBR=kerz

Review URL: http://codereview.chromium.org/3165031