DnsRRResolver: hoist inner class and callback on MessageLoop.
This change moves the Response class to the top level to save other
files from having to #include the header in order to get a pointer to
a Response object.

This change also modifies the semantics of resolution: the callback
will now be made of the origin MessageLoop, rather than a random
thread.

TEST=net_unittests
BUG=none

http://codereview.chromium.org/3110015/show