Merge 56403 - Handling empty ScopedDirLookup username in HandleDirectoryManagerEvent().
The constructor of syncable::ScopedDirLookup takes the username of the
account being synced. There are times when the username is blank,
resulting in a DCHECK being fired for the value of good() within the method
SyncManager::SyncInternal::InitialSyncEndedForAllEnabledTypes().

This patch makes sure that a ScopedDirLookup with a blank username is
never created by adding a check in the method
SyncManager::SyncInternal::HandleDirectoryManagerEvent().

BUG=51820
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/3184009

TBR=rsimha@chromium.org
Review URL: http://codereview.chromium.org/3170032