Merging r56866 to M6 branch: Original review: http://codereview.chromium.org/3112022
A crasher in TabContent::ExpireInfoBars is consistently reported in the field. Somehow one of the infobar delegate is NULL.
After spending some time investigating, I am still not clear on how this can happen.
Applying a band-aid with a NOTREACHED() in the hope someone will be able to repro consistently.

BUG=47669
TEST=Don't know how to repro :-(
TBR=kerz

Review URL: http://codereview.chromium.org/3186022