Revert 56966 - Port r56727 (agl@chromium.org) to the 472 branch.
*** Will re-apply if testing is successful ***

Differences on the 472 branch from trunk:
  -- HttpStreamRequest::DoInitConnection() is still in http_network_transaction.cc
  -- There's no ssl_config() getter, only the ssl_config_ member variable in http_network_transaction.cc
  -- There's no dnssec_enabled flag

____ original log message ____

net: expect MITM attacks with HTTP proxies and command line flag.

With r51258 we started requiring the TLS renegotiation extension from a
whitelist of servers that we knew supported it. When Chrome is getting MITM
attacked, this extension can be removed and this broke some debugging tools
(which intercept SSL connections) and some proxies which do the same.

This patch causes us to expect to be MITM attacked when tunneling via an HTTP
proxy and when the --allow-ssl-mitm-proxies command line flag is given.

BUG=48485
TEST=Can't really test without one of these MITM proxy machines.
Review URL: http://codereview.chromium.org/3125034

TBR=mal@chromium.org
Review URL: http://codereview.chromium.org/3191017