Remove attempt to be smart about where to open navigations targetting app tabs. I futzed with this a bit to try and integrate installed apps extents, but it is complex.
Also, I looked over the original bug where this code was
added: crbug.com/29281. The issue there was that when you
click on a bookmark (presumably in the bookmark bar) the
pinned tab is unexpectedly overwritten. I can see that
complaint, but there is also the use case of clicking
a bookmark and intending it to overwrite the pinned tab.

In summary, I kinda feel like the expectations about
how navigation should work are too ingrained and we
shouldn't be meddling.

BUG=52680

Review URL: http://codereview.chromium.org/3161037