Try to fix top crasher.
The problem is timing related so I am unable to write a unit test which proves this fixes things.  The essence of the change is to make sure we don't send messages to objects which no longer exist by clearing the delegate pointer in objects we own before we go away (where "we" is an object).  The delegate pattern does not get retained, which gives certain objects to hold dangling pointers.  

BUG=53205  http://crbug.com/53205

Review URL: http://codereview.chromium.org/3158031