Make SpeechInputManager a process wide singleton so it can make sure there is only 1 active audio recording happening at any time. Without this change, each SpeechInputDispatcherHost has it's own instance of SpeechInputManager so multiple tabs/windows can record audio simultaneously (and get the same audio data!).
With SpeechInputManager a singleton, it is no longer possible for the (render-view-id, request-id)
pair to be a unique identifier for a caller. We need to add render-process-id into the mix. None
of the speech input code cares about any of these and they can just work with an opaque identifier,
so I've added a SpeechInputDispatcherHost::SpeechInputCallers class which maps between the tuple of
3 values above to a single int/ID which can be passed as the caller id. With this change I have
removed the SpeechInputCallerId type replacing it with an int everywhere.

Also being a singleton the same SpeechInputManager is used by all SpeechInputDispatcherHost
instances, hence their delegate pointers could no longer be passed via the constructor of
SpeechInputManager. Instead we now pass the delegate to SpeechInputManager::StartRecognition and
it gets stored as part of the speech input session/request data. To implement this I have added
SpeechInputManager::SpeechInputRequest type and modified related code in that class accordingly.

BUG=none
TEST=All existing speech input unit tests and browser tests should pass. In manual testing, when recording is in progress in one browser window starting speech input in another browser window should abort/stop the first recording session.

Review URL: http://codereview.chromium.org/3116048