Jingle_glue bugfixes.
Fixed Closed() methods in JingleChannel and JingleClient so that they are
not blocking and guarantee that callback is not called afterwards.
Changed JingleThread shutdown mechanism so that the thread doesn't stop until
all tasks are finished.
JingleClient now uses LocalTunnelSessionClient that allows local connections.

BUG=52889
TEST=unittests

Review URL: http://codereview.chromium.org/3167047