Stop NetworkChangeNotifierMac from calling a virtual function before the constructor finishes.
NetworkChangeNotifierMac was posting a task to a separate thread in its constructor.  It's possible (although highly unlikely, but happens on buildbots) for the separate thread to start up and run the task (which invokes a virtual function) before the constructor completes (and, more importantly, the subtype's constructor initializes its vtable entries which are NULL / pure virtual in the base class).
The solution is to simply use a Delegate instead.  The Delegate should have been fully constructed (and thus, will not have the vtable initialization race) before being passed into NetworkChangeWatcherMac's constructor.  This also solves the stylistic issue of avoiding multiple inheritance, since NetworkConfigWatcherMac was not strictly an interface.

BUG=53138

Review URL: http://codereview.chromium.org/3344002