Revert 58601 -
A lot of ui_tests flakiness started around that revision, I want to verify if this is the cause.

original comment:

Have the browser start connecting as soon as the navigation starts.  This
seems trivial, but this gets the network going before we start the renderer
process creation, which can save 45+ms of latency.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/3358011

TBR=mbelshe@chromium.org
Review URL: http://codereview.chromium.org/3325014