[GTK] change workaround for weird RTL label alignment
When a GtkLabel has line wrapping and an integer size request, the text block is left aligned even in RTL locales. It will wrap at the left (which is correct) and be right justified (also correct) but is flush with the left of the widget allocation instead of the right of the widget allocation.

seems not to occur when the size is controled via gtk_label_set_width_chars() instead of gtk_widget_set_size_request(). (also goes away when line wrapping is turned off)

TODO: file a reduction upstream.

BUG=52857
TEST=launch chrome in Hebrew and look at the under the hood tab of options. Checkbox labels should look correct

Review URL: http://codereview.chromium.org/3355007