Revert 59123 - Don't use the return value of rlz_lib::SendFinancialPing() as the return value of the chrome extension API, since a false does not represent an error. Added an optional callback to the API so that caller can determine if the ping was sent or not.
This CL is to reapply change 58827, which was reverted due to a test break.
I found an uninitialized variable in the underlying RLZ code, see
http://code.google.com/p/rlz/source/detail?r=18 for the fix.  This CL includes
a bump of the RLZ library to get this fix.  The tests have been expanded a bit
too since 58827.

BUG=54294
TEST=n/a

Review URL: http://codereview.chromium.org/3350016

TBR=rogerta@google.com
Review URL: http://codereview.chromium.org/3330019