Lock fullscreen chrome on transition from one one bookmark folder to another.  Eliminate transition hazard.
BUG=http://crbug.com/53550

TEST=Place 2 bookmark folders on bar.  Go fullscreen.  Move mouse up
so chrome opens.  Click on one folder, then scroll down a little.
Move mouse to the side (so it's in a position that would NOT trigger
the chrome to be open) then quickly go up and click on 2nd folder.  If
the 2nd folder has the right y position, success!  If it's pegged to
the top of the screen, fail.  Repeat a few times.




Review URL: http://codereview.chromium.org/3297024