Linux zygote: Fix race condition when using SUID + seccomp sandboxes
Change the child process to synchronise with the parent (the zygote)
so that the child does not fork further children while the zygote is
trying to discover the child's real PID from the SUID helper.

Split the logic for doing METHOD_GET_CHILD_WITH_INODE into a separate
function.  Add a pipe for doing the synchronisation.

Now that the child synchronises, it can close dummy_fd.

BUG=55599
TEST=Tested manually with:
  cd native_client && ./tools/httpd.py &
  CHROME_DEVEL_SANDBOX=/opt/google/chrome/chrome-sandbox ./out/Debug/chrome-wrapper --enable-nacl --enable-seccomp-sandbox http://localhost:5103/tests/prebuilt/srpc_hw.html

Review URL: http://codereview.chromium.org/3391003