Merge 59794 - Fix case where we close a stream due to socket errors when it is currently a pending create stream.
The problem was that the socket error would arrive on the SpdySession, which
would initiate cleanup of the pendng create streams for that session.  In the
callback to the stream, the stream would notify the transaction, which would
delete itself, and that would cause a callback into the SpdySession to
clear that specific pending create stream (via CancelPendingCreateStreams).

BUG=52901,55795
TEST=SpdyNetworkTransactionTest.ThreeGetsWithMaxConcurrentSocketClose

Review URL: http://codereview.chromium.org/3400009

TBR=mbelshe@chromium.org
Review URL: http://codereview.chromium.org/3447011