Split dom_perf into individual tests.
dom_perf began timing out recently because running
all of its tests was taking longer than the Buildbot
hang-on-no-output timeout.  We split dom_perf calls
into running individual tests to avoid that.

This change requires computing the geometric mean of
results which was done by the dom_perf benchmark before.
Since the benchmark instance no longer has access to
all of the results for a particular configuration, we
do this in the Python harness, instead.

Fix the return code logic to allow failure when just
one of the tests fails.

BUG=55540
TEST=dom_perf runs as before, updates perf dashboard
as expected
Review URL: http://codereview.chromium.org/3465001