Revert 60082 - Add a PhishingClassifier implementation.
The PhishingClassifier ties together the feature extractors and the scorer,
and will be invoked by the RenderView when a page finishes loading.  This
change also makes the feature extractors have a non-owned pointer to the clock
so that it can be shared, and fixes a problem where the DOM feature extractor
cleared out any already-present features.

BUG=none
TEST=PhishingClassifierTest

Review URL: http://codereview.chromium.org/3452010

TBR=bryner@chromium.org
Review URL: http://codereview.chromium.org/3436026