Add some optimizations to plugin painting.
The simplest one is to disable blending when the plugin is opaque.

The more complicated one is to bypass webkit painting the background of plugins
when we know the plugin to be always on top and also opaque. The always on top
flag is currently set by a new "Private2" API. Bypassing WebKit makes
animations faster.

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/3421030