Merge 59711 - Disk cache: Fix the order in which we delete data from the block files.
Stress testing the cache reveals a problem with the deletion
of some data from an entry: it is possible to crash in a way
that the block file thinks a block is free and an entry
thinks the block is in use. This CL corrects that issue.

There is also some new tests and a bunch of DCHECKS added
while looking for the problem, as well as adding tests to
make sure that a block file is accessed only from one thread
(there is no problem with the current code in that regard)

BUG=55605
TEST=netunittests

Review URL: http://codereview.chromium.org/3430004

TBR=rvargas@google.com
Review URL: http://codereview.chromium.org/3389030