Remove chrome_frame*.pdb from the archive.
Right now the windows slaves throw an MemoryError exception when they try to
decompress the zip. This patch reduces the size of the zip so we can get it
going for tonight. This should be revisited to more finely choose which pdb to
keep.

TEST=none
BUG=none

Review URL: http://codereview.chromium.org/3388020