Improve flaky ui_test AutomationProxyVisibleTest.AutocompleteMatchesTest
The omnibox requires that it is focused before text input to it will
kick off the autocompletion providers. We add code
(edit->WaitForFocus()) to the test to wait for the focus to be obtained
before setting the omnibox’s text input and proceeding with the rest of
the test.

BUG=19876
TEST=Existing

Review URL: http://codereview.chromium.org/3348021