browser_tests: dial back test timeout from 45 to 20 seconds.
In this bug
  http://code.google.com/p/chromium/issues/detail?id=43862
we upped it to 45 because a test was timing out.  That test is
currently taking 4.6s.

If we have any tests that require more than 20 seconds to run,
we should adjust the particular test to require a longer timeout.

TEST=tests like SSLUITest.TestUnsafeContents, which are timing out,
no longer take 45 seconds to time out.

Review URL: http://codereview.chromium.org/3466020