Remove spammy warning about reaching maximum socket count for the pool.
When I run Chrome through a proxy, this WARNING log triggers very frequently. Typically immediately after it's called, CloseOneIdleSocket() is invoked and a connect job is still allowed to proceed.

If this WARNING was useful in the past, I will add it to the call sites at appropriate places (such as when there are no idle sockets for the pool and a request must be queued) rather than inside ReachedMaxSocketsLimit().

BUG=None
TEST=None

Review URL: http://codereview.chromium.org/3526013