Simple fix correcting the lifetime scope of automated url request jobs during chrome-initiated xmlhttprequest aborts.  Two similar problems are fixed by this CL.
I found a race condition during the early-termination processing of the NpapiUrlRequest objects.  If NpapiUrlRequest::Stop is invoked before NpapiUrlRequest::OnStreamCreated, then it is possible that a new stream will be created on the Chrome-Frame side, for which an operation that has been officially cancelled within Chrome.  The fix is to add a new parameter fo the NPAPIUrlRequest, stop_requested_ that is used to track this behaviour.  If a new stream is created on a request that has been stopped, then that stream is immediately destroyed, without notification back to Chrome.

If Chrome requests an abort of a url request, then Chrome-Frame should not send any notifications back to Chrome concerning the tear-down of that request.  The delegate is revoked in NPAPIUrlRequest::Stop to achieve this behaviour.

BUG=None
TEST=None

Review URL: http://codereview.chromium.org/3530002