Fix transaction rollbacks in web_database migration code
Changes |WebDatabase::MigrateOldVersionsAsNeeded()| to return error status should database migration fail.  If migration does fail then the initialization as a whole is rolled back.  The error code is returned to caller so user is alerted.  This also cleans up logging so that in all error cases log warnings are emitted.  Additionally, this reworks the Autofill credit card migration logic to test the schema before migrating.  This is preferable to the existing logic that attempted the migration and then looked for failure after the fact.

BUG=54012
TEST=WebDatabaseMigrationTest.*

Review URL: http://codereview.chromium.org/3615007