DOMUI Options: Fix vertical alignment of radio and checkboxes
This also makes the label text line up correctly if the label needs to wrap.

This also fixes a bug where the menu could be scrolled off the screen.

I went through and fixed a bunch of indentation errors in the HTML I touched.

BUG=None
TEST=None

Review URL: http://codereview.chromium.org/3590006