git_cl_hooks: get cl fields vi git_cl accessors instead of using backquote
This is part 2 of a 2-part change which fixes a bug where git-cl
hangs if ~/.codereview_upload_cookies does not exist. It is
actually waiting on user input for email/password (but you don't see the
prompt because it's happening in a subprocess).

This is a do-over of http://codereview.chromium.org/3622002.

Much thanks to Lei Zhang for fixing the bugs in the original CL. His
fixup CL was http://codereview.chromium.org/3541019.

BUG=none
TEST=Verified that the hang is fixed.

Wrote a quick pre-submit to test that issue and patchset are ints:
    http://codereview.chromium.org/3533020

Output:

** Presubmit Warnings **
Patchset is 1

Issue is 3533020

Review URL: http://codereview.chromium.org/3591017