Fix the default timezone detection in ICU.
This is the second patch for this issue. The previous patch was incomplete
in that if a match was found for /etc/locatime in 'posix' directory under
/usr/share/zoneinfo (say, posix/America/New_York)
instead of at the top-level in /usr/share/zoneinfo (say, America/New_York),
the DST offset is set to 0 while the raw-offset is set correctly.

This leads the history and file:// to show the time stamp one hour behind
the 'wall-clock time' when the DST is in effect.

The upstream bug is http://bugs.icu-project.org/trac/ticket/8014


BUG=38632
TEST=1. Make a file locally and open the directory with the file when
daylight saving time is in effect. The time stamp should match that of
'ls' command.
2. Visit a web page and open up the history tab and see the entry has
the correct time stamp instead of time stamp 1-hr behind when daylight
saving time is in effect.

Review URL: http://codereview.chromium.org/3598015