Fix inconsistent line endings problem.
For some reason invoking 'svn cat' on this file was failing with an error
about inconsistent line endings, and that is preventing me from using gcl
upload of a review where I have some changes. I just ran 'dos2unix' on the file
for this CL - hopefully that will fix the problem.

BUG=none
TEST=none

TBR=tim@chromium.org