Merge 62524 - Fix a drag-and-drop bug in the bookmark manager on Chrome OS.
Before the fix, drag-and-drop in the bookmark manager on Chrome OS
resulted in a copy rather than move. The drag-and-drop logic is shared
with GTK and Views, but there was a subtle difference between the two
toolkits, that caused the bug.

BUG=chromium-os:7533
TEST=confirmed that dnd in the bookmark manager worked on Linux and Chrome OS

Review URL: http://codereview.chromium.org/3743002

TBR=satorux@chromium.org
Review URL: http://codereview.chromium.org/3747004