Tweak the webstorePrivate.getBrowserLogin callback signature.
I accidentally defined it as taking 2 parameters intstead of a single object
with two named properties. This CL fixes that (and incidentally happens to make
it match the callback for the promptBrowserLogin callback, which is nice for
consistency).

I also did a little cleanup and added a test for getBrowserLogin and added 
another test for promptBrowserLogin while I was in the neighborhood.

BUG=58033
TEST=none

Review URL: http://codereview.chromium.org/3767010