Add an option to the test_shell to enable it to load each test several times
We have sees bugs which could have been caught by reloading a page one or more times and finding that the results for one of the subsequent loads is not the same as the expected output. The new option--multiple-loads=X will load each test X time. The output dumped by test_shell is the result of the last load. The default when the option is specified without a value is 2 times in debug mode and 5 times in release mode.

To be able to have more fine-grained control of how the JavaScript engine behaves for each load the existing flag --js-flags can now specify a list of flag-sets like this

  --js-flags="--xxx,--noxxx --yyy,--noyyy"

First load will run with --xxx, the second with --yyy and the third without any (the 'no' prefix is handled by V8 to turn off the flag).

The changes to the Python test runner will be in a separate change.

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/4016002