Take 3: Refactor, relocate and rename ProfileSyncServiceTestHarness.
ProfileSyncServiceTestHarness is an automation helper class that lives in chrome/test. It contains methods that allow you to wait on various sync events, and is used by the sync integration tests.

We now need to re-use ProfileSyncServiceTestHarness to support sync operations in TestingAutomationProvider, a class that lives in chrome/browser. In order to accomplish this, we do the following things in this patch:

1) Move ProfileSyncServiceTestHarness to chrome/browser, so that TestingAutomationProvider can re-use its functionality, and rename it to ProfileSyncServiceHarness (a less wordy name that reflects the purpose of the class).
2) Since TestingAutomationProvider relies on the fact that the UI message loop is always running, we must change the default behavior of ProfileSyncServiceHarness to not start/stop the UI message loop. We make 2 key wait methods pure virtual: SignalStateComplete() and AwaitStatusChange().
3) The integration tests use a specialization of ProfileSyncServiceHarness that start/stop the UI message loop.
4) TestingAutomationProvider can go on to use a specialization of ProfileSyncServiceHarness that uses WaitableEvents. (coming up in another patch)

NOTE: There were 2 earlier attempts at checking in this patch -- See
http://codereview.chromium.org/3492005/show and
http://codereview.chromium.org/3419029/show.

BUG=56460
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/4145003