Forward textfield focus event to the browser.
Add two messages in IPC channel: 1. The renderer needs to inform the browser if an element that can accept use input gains focus. This is useful when the browser may want to take some action against such events. For example, in case a textfield is focused, the browser may want to show an on-screen keyboard. 2. The browser needs some way to inform the renderer to scroll the currently focused element into the view frame. This is useful when, for instance, the browser resize the frame and the focused element goes out of visible frame.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/3474007
Patch from Varun Jain <varunjain@chromium.org>.