Adds error text for the ERR_NETWORK_ACCESS_DENIED error, which is generally caused by a firewall.
Also removes the title strings for proxy and disconnected
errors, replacing them with pre-existing generic ones, and
sorting them by their new title.  This avoids a debug
assertion caused by the fact neither title includes the url
of the original page.  Not much of titles are seen anyways,
so not worth the effort of a workaround.

BUG=57108
TEST=manual

Review URL: http://codereview.chromium.org/4182012