Var serialization-related proxy stuff. This allows vars to be sent over IPC and received with proper refcounting semantics.
The basic design is that the SerializedVar is the thing actually handled by
IPC. Then you use one of the many helper classes depending on whether you are
sending or receiving and what the ownership semantics are. The helper classes,
along with the VarSerialization interface makes the right thing happen for
string conversion & refcounting.

The VarSerialization class is implemented differently on the browser and plugin
side and is where the differences in refcounting & strings are handled. This
allows the rest of the code to be the same on both the plugin and browser side
and allows vars to be sent back and forth.

This patch references some files not in it like the dispatcher and various
tracker classes. This is the cleanest cut I could make for a reasonably
reviewable chunk.
Review URL: http://codereview.chromium.org/4096008