Moved --use-gl=osmesa switch out of test_launcher_utils.
It was affecting tests other than ui_tests and browser_tests, which potentially need to use regular GL. For example, plugins expect to be able to use regular GL with core animation.

In the future, we should be able to make more tests use osmesa by allowing a regular GL surface to be read back into an OSMesa surface in AcceleratedSurface on Mac.

TEST=ui_tests, browser_tests
BUG=62602

Review URL: http://codereview.chromium.org/4761001