Make the Graphics2D flush call an asynchronous operation. This tracks the callback in the plugin and adds a new message when a flush is complete in the renderer.
I adapted the callback factory so it could be used for tracking the flush
callbacks in the proxy in the renderer, and also changed the URL loader
stuff (which had some custom management) around so that it also uses the
factory.

This also fixes a bug in font proxying where we would assert if there was a
font creation error because the dispatcher wasn't set on the SerializedVar in
the font decription.

TEST=none
BUG=none
Review URL: http://codereview.chromium.org/4876003