Mac: Let out-of-process tests run in bundled mode for their whole lifetime.
Since out-of-process tests override the EXE path to look like the bundled app, it makes sense to override AmIBundled() as well.

This is important because the renderer process started from browser_tests runs as bundled, and if browser and renderer process don't agree on bundled-ness, the "load plugin" requests for internal plugins from the renderer have the wrong plugin path, causing the plugin load to fail.

Also add a DCHECK that makes sure that AmIBundled() doesn't flip-flop.

This makes PDFBrowserTest work on mac, so enable it.

It looks like even unit_tests uses the out-of-process test runner, so this change is a bit hairy :-/

BUG=61258,63183
TEST=all existing tests still pass, PDFBrowserTest.* passes.

Review URL: http://codereview.chromium.org/4947002