Merge 65840 - Provide sync integration tests with a way to set a favicon for a bookmark.
Note: This is required to fix the sync tests on chrome OS.
See http://code.google.com/p/chromium-os/issues/detail?id=9262

The sync integration test framework does not provide tests with a way to
set a favicon for a bookmark and verify that it gets synced.

This patch adds a method called SetFavicon, that allows tests to set
either a generic favicon, or a variable favicon (using an index) for a
bookmark.

BUG=57234
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/4749003

TBR=rsimha@chromium.org
Review URL: http://codereview.chromium.org/5149001