Convert the traces in the WhitelistedExtension test.
The test has been enabled for a long time and does not
currently even appear on the Flakiness dashboard, nor can I
find a failure on the try servers.

The failure mentioned in the bug does not show the traces I
added, so I'm converting the traces to another form, in
case we see flakiness again.

BUG=57694
TEST=Changing a test.

Review URL: http://codereview.chromium.org/5255004