Merge 67720 - chromeos: Update static about:os-credits page.
The old version was way out of date.  I generated a new one
using a script (not sure where to check it in) that goes
through all of the packages installed for a board, fetches
the tarball for each and looks for a license file in it, and
then falls back on the LICENSE variable from the ebuild
file.  There's a lot of one-off tweaking too. :-(

BUG=chromium-os:7840
TEST=looked at it

Review URL: http://codereview.chromium.org/5357010

TBR=derat@chromium.org
Review URL: http://codereview.chromium.org/5281014