Handle shift-selection in GroupTables.
Change the shift-click support in TableView so shift-clicks that span multiple
groups are prohibited. Clicking in one group, then shift-clicking in another
group will cause the shift-click to be treated as a normal click (i.e. just
select the item in the second group).

BUG=63787
TEST=Open task manager, click on item in one group, shift click on item in another group.

Review URL: http://codereview.chromium.org/5504001