Use stopped_on_render_loop_ to prevent further work from executing in BufferedDataSource.
Previously we were using it as a DCHECK but we would fail gracefully.  Now that stopped_on_render_loop_ is set via Abort() (i.e., catastrophic failure), we need to use it as a singal to make sure all work on the render loop does indeed stop.

BUG=16751
TEST=http/tests/media should stop crashing

Review URL: http://codereview.chromium.org/5603004