Modify the "dangerous download" algorithm as follows.  Original patch by Pierre-Antoine LaFayette (see http://codereview.chromium.org/1403001/ ), r=me,brettw,aa.
Downloads are considered dangerous if: 
a) The file is dangerous just by sitting on the drive, without needing to be clicked on e.g. dll, xbap 
b) The file is executable and the download was not user initiated. 
c) They are an extension that is not from the gallery 

We have defined a user initiated download as 3 possible cases: 
a) A user enters a URL into the address bar that is a file 
b) A user left clicks on a URL that is a file 
c) A user right clicks and does "Save As" on a URL that is a file. 

BUG=9044
TEST=Open a page with a download link to a dangerous file that is not an extension, e.g. an .exe file, and left click on the link. The download should proceed without a prompt.
Review URL: http://codereview.chromium.org/5603008