Disabled GPU watchdog thread on Windows.
I coulndn't find good enough heuristics for automatically aborting the GPU process on Windows. With this change we still have the browser's watchdog, which prompts the user if they want to kill the process if it times out.

The GPU watchdog timer seems to be working okay on Mac and Linux so I have left it enabled. For the same reason, I took out the field trial. Previously, only 15% of users were in the field trial. Now they all are.

TEST=about:gpuhang, try
BUG=65229,58396,64655

Review URL: http://codereview.chromium.org/5537006