Merge r68241 from the 552 branch
Fix case where a SPDY stream with an un-verified cert would kill the browser
if an https:// URL was attempted to be fetched across it. Also fix a bug
where when two SSL connections are made, but only one is moved into a SPDY
session (because the second is redundant), close the redundant connection.
This had been leaking a un-verified SSL connection into the pool, which
could then get re-used as though it was verified.

BUG=64861
TEST=none