[Chrome OS] Add README to explain wonky DEPS file use
We are moving to using a DEPS file for extra chromium os deps that is checked into the main Chromium tree.  As an artifact of the way gclient works, this means that devs will have a r/w version of this file in their source tree and _also_ a read-only version that gclient checks out somewhere else.  This README explains this and how to update the DEPS file correctly.

BUG=None
TEST=None

Review URL: http://codereview.chromium.org/5587011