Fix flaky BrowserKeyEventsTest.ReservedAccelerators on Windows (hopefully)
According to the log, the flakiness may be caused by http://crbug.com/65375. But I can't reproduce it manually on my local Windows 7 workstation. This CL just uses the same test flow as Mac port to try to avoid flakiness. The FLAKY label is not removed in case the test will still be flaky.

BUG=65847
TEST=The test should not flaky on Windows.

Review URL: http://codereview.chromium.org/5746002