Repoint the ESET error Learn More link.
The ESET error occurs when Chrome's HTTPS connections are intercepted by
ESET. However, the Learn More link on the error page points to an HTTP
URL which redirects to an HTTPS URL. Since the user cannot view HTTPS
pages, that's unhelpful.

This change points the link to a page on ESET's knowledge base instead.

http://codereview.chromium.org/5705003/