Fix pyauto downloads tests to trigger dangerous downloads to match expectation
It used to be possible to navigate to a dnagerous file type to trigger its
download. Not anymore. Now all downloads where the user initiated it (by
navigating to a URL or clicking on a link) are not marked unsafe even for
unsafe filetypes.

Update the tests to trigger dangerous downloads by navigating to a file which
in turn triggers the dangerous files to download.

BUG=66081, 57875

Review URL: http://codereview.chromium.org/5781006