Factory identifiers have been created years ago when we needed to pass only one bit of information around in the factory. Now we have much bigger needs and it has been replaced by the factory_properties.
However, we still continue to pass around the factory identifier, and use
it at a few places (mostly incorrectly).

This change removes the identifier completely, so it cannot be misused and
to remove the complexity that it adds.

Some of the places where we were using it have been converted to factory_properties.

Review URL: http://codereview.chromium.org/5863003