DOMUI Prefs: Add support for multiple levels of sub-pages
The JS is mostly max-depth-agnostic because it's cleaner that way. The HTML has the two layers hard-coded because it's less convoluted than creating the layers dynamically and moving things, there's no expectation at the moment that we will have more than 2 levels of subpages. If it looks like that might happen later on we can revisit how we construct layers.

BUG=None
TEST=None yet; actual sub-sub-pages will be added soon.

Review URL: http://codereview.chromium.org/6024001