Ensure that chrome frame tests don't hang indefinitely if IE does not respond to outgoing COM calls issued by these tests. These tests hang intermittently on the IE9 builder on the experimental waterfall.
To achieve this we register a COM message filter and set a windows timer in the ChromeFrame
TimedMsgLoop object. In the message filter we look for a WM_TIMER created by the message loop
object and on finding it abort the outgoing COM call.

This flag is percolated to the related event sinks which leak the underlying interfaces to
prevent the Release calls from hanging too.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/6051003