Support for GCF ready-mode.
The installer can now be run in a way that installs both Chrome and Chrome Frame
at the same time, into a shared directory.  The command line to do this is:

mini_installer.exe --multi-install --chrome --chrome-frame --ready-mode
--system-level --verbose-logging

This installs both products although only Chrome will have an entry in the
add/remove programs dialog that serves as an uninstallation command for both
products.  Chrome Frame will be installed in ready-mode, which means that the
user will have to opt-in or opt-out of using it.  The installer will create a
REG_DWORD value to indicate this mode, here:


HKEY_LOCAL_MACHINE\SOFTWARE\Google\Update\ClientState\{4DC8B4CA-1BDA-483e-B5FA-D3C12E15B62D}
 Value name: ChromeFrameReadyMode
 Value: 1

If the user opts-in to use GCF, setup will be run again to remove this value,
update Chrome's uninstallation commands to only uninstall Chrome, and add an
entry to the Add/Remove Programs dialog for GCF.  To do this, the installer
needs to be run with this switch:

 --multi-install --chrome-frame --system-level --ready-mode-opt-in --verbose-logging

If the user opts-out, Chrome Frame will be uninstalled by running the installer
with these arguments:

 --uninstall --chrome-frame --ready-mode --system-level --multi-install --verbose-logging

In addition to uninstalling GCF, this updates Chrome's uninstallation commands
accordingly and sets the ChromeFrameReadyMode value to 0 to avoid enabling ready
mode again.

Requirements that must currently be met in order to install Chrome and Chrome
Frame into the same shared location:
- Chrome Frame must not already be installed
OR
- Chrome Frame must be already installed as multi, into the shared location

**** other changes ****

The installer does no use the ap value for detecting other multi-install
products anymore.  Instead we inspect the uninstallation switches to see if
--multi-install is present.  Updated unit tests accordingly.

When uninstalling Chrome Frame along with Chrome, the Chrome Frame
uninstallation prompt is not shown.

Added safeguards:
If we attempt to install Chrome Frame multi over a single, install aborts and returns an error.
If we attempt to install Chrome Frame single over a multi, install aborts and returns an error.

TEST=See description above.
BUG=61609

Review URL: http://codereview.chromium.org/5989007