Increasing live (sync) operation timeout to 45 seconds.
The sync integration tests were using a timeout of 30 seconds for live
operations. Some of the sync stress tests are timing out because it
takes longer for a full sync cycle to complete.

This patch increases the timeout to 45 seconds, and should eliminate the
occasional failures on the builders.

TBR=tim@chromium.org
BUG=68207
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/6023011