Autofill heuristics regular expressions should be stored in external data files.
This change moves the Autofill heuristics regular expressions out to a new autofill_resources.grd file.  This grd file is not meant to be localized.  Eventually it will contain the locale-specific regular expressions as well as the English versions.  This CL, however, just moves the English versions out of source.

BUG=55787
TEST=FormStructureBrowserTest.HTMLFiles

Review URL: http://codereview.chromium.org/6026010