Enable "inspect element" for chrome:// pages.
It's already possible to open the js debugger for these pages via the wrench menu, so disabling "inspect element" is just a nuisance. Also, there's no obvious reason in my mind why it should be different for debug/release (I do most of my domui development in release as it's faster to build and easier to use the debugger).

BUG=none
TEST=manual

Review URL: http://codereview.chromium.org/6094007