Enable accelerated compositing in fullscreen pepper.
The compositing itself is trivial since there is just a fullscreen plugin instance (that we blit full-screen), but it allows the use of the 3D contexts in pepper when fullscreen is on (they require accelerated compositing).

Accelerated compositing is only triggered when a Surface3D is bound to the instance. Otherwise the regular 2D painting path is used.

This CL also refactors a bit RenderWidgetFullscreenPepper, merging the trivial WidgetFullscreenContainer into RenderWidgetFullscreenPepper - it's simpler that way.

BUG=none
TEST=Flash

Review URL: http://codereview.chromium.org/5994002