Add methods GetWindowBounds and IsWindowMaximized to TestingAutomationProvider.
I'm adding a test using these methods to another larger patch, so I want to break this file out so as not to encumber the big patch.  Test using these will be in that bigger patch (hard to break that test out nicely because it's part of a wholescale preferences migration).

BUG=none
TEST=next patch
Review URL: http://codereview.chromium.org/6124006