Fix null pointer dereferences in IndexedDB
When the channel is disconnected, the various dispatcher hosts are released
and sent to the WebKit thread for dereferencing.  Between then and when the
object actually gets destroyed (which could be some time later), it's possible
for map_->Add() to be called on the various dispatcher hosts.  Clearly this
can lead to a null dereference.  Thus we should delete the object and return
zero in such cases.

BUG=68078
TEST=none

Review URL: http://codereview.chromium.org/6172004