give tabbed options clear browsing dialog a facelift
It should now match the mocks modulo the adobe flash settings link (not sure why the mocks show it as a checkbox; AFAIK making it a checkbox is outside our ability to do).

also make it work (the time period select dropdown was broken).

BUG=69163,63843
TEST=manual

Review URL: http://codereview.chromium.org/6110006