Add support in the ChromeFrame NPAPI plugin for receiving URL redirect notifications. This is basically implementing the NPP_URLRedirectNotify plugin function and invoking the browser end function NPN_URLRedirectResponse indicating whether the redirect is to be followed or not.
The ChromeFrame NPAPI implementation always disallows url redirects from the plugin and
instead expects Chrome to follow the redirect.

Tested this with a Firefox 4 nightly build. Currently this does not work as expected because
of a bug in Firefox where it invokes the NPP_URLRedirectNotify  function with the original
URL instead of the redirected URL. This is tracked by mozilla bug https://bugzilla.mozilla.org/show_bug.cgi?id=625164

BUG=69419
TEST=none. Will add a test for this once we have a builder with a working version of Firefox 4.


Review URL: http://codereview.chromium.org/6223003