Update buildbot command line parameters passed to the sync integration tests
In the past, the sync integration tests read GAIA credentials from a password file on the chromium bots. With revision http://src.chromium.org/viewvc/chrome?view=rev&revision=71253, this is no longer necessary.

In addition, some tests used to need up to 5 minutes to complete due to several clients syncing to a server. As of today, the tests have been re-designed to be self contained with no external server dependencies, and as a result, are much faster.

This patch does 2 things:
- Gets rid of the --password-file-for-test parameter.
- Reduces the timeout value passed to --test-terminate-timeout from 300000 (5 minutes) to 120000 (2 minutes).

BUG=69578,69586
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/6274004