Sync: Clean up fake URLFetcher factory prior to browser shut down
Revision 71253 introduced a fake URLFetcher factory to the sync tests.
The clean up step of reverting to the default URLFetcher was added to
LiveSyncTest::TearDown(), which is run after browser shut down.

The correct place to add the clean up step is the method
LiveSyncTest::TearDownInProcessBrowserTestFixture(), which is run prior
to browser shut down. This patch merely moves the clean up step to the
right spot.

TBR=akalin@chromium.org
BUG=69184, 56729, 53931
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/6251004