tabbed options - new clear browsing data overlay
I tried to make this look as nice as possible (and close to the mocks), but Alex will have to fiddle with the CSS to finalize it. I didn't implement the button changing to "clearing..." because it changes the width of the button and thus the whole overlay. The throbber + disablement should hopefully be enough of a visual cue.

BUG=63843
TEST=manual

Review URL: http://codereview.chromium.org/6340002