DOMUI Prefs: Add inline buttons for setting the default search engine
This is an interim solution for M10 since the final approach (drag and drop) will likely not be ready for M10. This adds inline buttons on hover/edit in the same style as the password field "Show" button, which gives a far better experience (especially when there are a lot of engines) than the old model of a single button on the side.

BUG=63825
TEST=Hovering or editing a search engine should show a "Make Default" button.

Review URL: http://codereview.chromium.org/6352004