Make the code for unmaximizing the window during SetBounds much simpler.
This still fixes the original bug and does it in a much simpler (and faster)
way. I was previously attempting to use an API to do this rather than just
setting the style directly but that API may have hidden side effects like
causing a window activation on some versions of Windows.

BUG=69724
TEST=WidgetWinTest::SetBoundsForZoomedWindow

Review URL: http://codereview.chromium.org/6331006