Initialized pair_id variable to 0.
This variable was uninitialized and the compiler gave a warning (which
got converted into an error because we compile using -Werror) when
GetIDAndCountOfFormElement() and its children are inlined. This is
because there is a path in that function that returns true but leaves
the pair_id variable uninitialized (if s.Step() is 0).

Note that I am not a Chromium committer so you will have to commit
this.

Patch by asharif; http://codereview.chromium.org/6293011/

BUG=70144
TEST=no visible change