ppapi_egl target is currently unused as verified by using:   grep --include="*.gyp*" -nr -F "ppapi_egl" .
However this library causes NaCl to require third_party/mesa which relative to NaCl is large.

This CL moves the definition of this static library from ppapi to the lib/gl directory under ppapi so that PPAPI users are not forced to DEPS in mesa.

Review URL: http://codereview.chromium.org/6342012