Observe if MailNotifier subscribes to this builder.
MailNotifier takes a 'categories' arg that is later used
in MailNotifier.builderAdded to indicate whether or not
to subscribe to a builder.  ChromiumNotifier overrode
builderAdded yet didn't re-implement that categories
check, so we would end up subscribing to a builder even
if that builder's category didn't match what was passed
to us.  The result would be an extra email (since we're
sending mail for all of the builders we want plus all
the builders we don't want).

Add a call to the parent class's builderAdded first and
return None when that call returns None.  This should
re-implement the same tests in the base class.

categories is used in MailNotifier.buildFinished,
however we completely override that class to avoid
sending mail when a build is complete.  The remaining
use of categories is in __init__ which we call in
ChromiumNotifier's __init__.

BUG=none
TEST=ChromiumNotifier observes the categories arg
Review URL: http://codereview.chromium.org/6084011